Moved checkbox & column number into routePath stop form
